requirements.txt
requests
pandas
matplotlib
python-telegram-bot==13.15
schedule
